Fix portability issues.

+++ cap-primitives/src/fs/manually/open.rs
@@ -531,7 +531,7 @@ pub(crate) fn stat(start: &fs::File, path: &Path, follow: FollowSymlinks) -> io:
 /// potentially being a symlink we need to follow, due to use of `O_PATH`.
 #[cfg(any(target_os = "android", target_os = "linux"))]
 fn should_emulate_o_path(use_options: &OpenOptions) -> bool {
-    (use_options.ext.custom_flags & OFlags::PATH.bits()) == OFlags::PATH.bits()
+    (use_options.ext.custom_flags & (OFlags::PATH.bits() as i32)) == (OFlags::PATH.bits() as i32)
         && use_options.follow == FollowSymlinks::Yes
 }
